requirements.txt
PyYAML>=5.4
pygame>=2.1
colorama>=0.4.6
kivy>=2.1.0
